MOving everything in to filter types
